@@ -14,70 +14,67 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/fire": "^6.0.0-rc.2",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
-    "@capacitor/android": "^2.0.1",
+    "@angular/common": "~11.0.5",
+    "@angular/core": "~11.0.5",
+    "@angular/forms": "~11.0.5",
+    "@angular/platform-browser": "~11.0.5",
+    "@angular/platform-browser-dynamic": "~11.0.5",
+    "@angular/router": "~11.0.5",
+    "@capacitor/android": "^2.4.5",
     "@capacitor/core": "2.0.0",
-    "@ionic-native/core": "^5.0.7",
-    "@ionic-native/splash-screen": "^5.0.0",
-    "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^5.0.0",
-    "@ionic/storage": "^2.2.0",
-    "@ngrx/effects": "^9.1.2",
-    "@ngrx/store": "^8.6.0",
-    "@ngrx/store-devtools": "^9.1.2",
+    "@ionic-native/core": "^5.30.0",
+    "@ionic-native/splash-screen": "^5.30.0",
+    "@ionic-native/status-bar": "^5.30.0",
+    "@ionic/angular": "^5.5.2",
+    "@ionic/storage": "^2.3.1",
+    "@ngx-translate/core": "^13.0.0",
     "cordova-android": "8.1.0",
-    "cordova-sqlite-storage": "^5.0.0",
-    "core-js": "^2.5.4",
+    "cordova-sqlite-storage": "^5.1.0",
+    "core-js": "^2.6.12",
     "express": "^4.17.1",
-    "firebase": "^7.14.4",
-    "moment": "^2.24.0",
+    "moment": "^2.29.1",
     "ngx-color-picker": "^9.1.0",
     "randomcolor": "^0.5.4",
-    "rxjs": "~6.5.1",
-    "tslib": "^1.9.0",
-    "uuid": "^8.1.0",
-    "zone.js": "~0.9.1"
+    "rxjs": "~6.6.3",
+    "tslib": "^1.14.1",
+    "uuid": "^8.3.2",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.20",
-    "@angular/cli": "~8.3.23",
-    "@angular/compiler": "~8.2.14",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
+    "@angular-devkit/architect": "~0.900",
+    "@angular-devkit/build-angular": "~0.1100.5",
+    "@angular/cli": "~11.0.5",
+    "@angular/compiler": "~11.0.5",
+    "@angular/compiler-cli": "~11.0.5",
+    "@angular/language-service": "~11.0.5",
     "@capacitor/cli": "2.0.0",
-    "@ionic/angular-toolkit": "^2.1.1",
+    "@ionic/angular-toolkit": "^2.3.3",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
     "codelyzer": "^5.0.0",
-    "cordova-plugin-device": "^2.0.2",
+    "cordova-plugin-device": "^2.0.3",
     "cordova-plugin-ionic-keyboard": "^2.2.0",
-    "cordova-plugin-ionic-webview": "^4.2.0",
-    "cordova-plugin-splashscreen": "^5.0.2",
-    "cordova-plugin-statusbar": "^2.4.2",
-    "cordova-plugin-whitelist": "^1.3.3",
+    "cordova-plugin-ionic-webview": "^4.2.1",
+    "cordova-plugin-splashscreen": "^5.0.4",
+    "cordova-plugin-statusbar": "^2.4.3",
+    "cordova-plugin-whitelist": "^1.3.4",
+    "firebase-tools": "^7.12.1",
+    "fuzzy": "^0.1.3",
+    "inquirer": "^6.5.2",
+    "inquirer-autocomplete-prompt": "^1.3.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
+    "karma": "~5.1.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "open": "^7.0.3",
+    "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.4.3",
-    "@angular-devkit/architect": "~0.900",
-    "firebase-tools": "^7.12.1",
-    "fuzzy": "^0.1.3",
-    "inquirer": "^6.2.2",
-    "inquirer-autocomplete-prompt": "^1.0.1"
+    "tslint": "~6.1.3",
+    "typescript": "~4.0.5"
   },
   "description": "A simple yet beautiful todo application",
   "cordova": {
