@@ -1,9 +1,7 @@
-import { AppService } from 'src/app/state/app.service';
 import { ModalModule } from './../shared/modals/modal.module';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouteReuseStrategy } from '@angular/router';
-import { EffectsModule } from '@ngrx/effects';
 
 import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
 import { SplashScreen } from '@ionic-native/splash-screen/ngx';
@@ -12,18 +10,9 @@ import { StatusBar } from '@ionic-native/status-bar/ngx';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HttpClientModule } from '@angular/common/http';
-import { AngularFireModule } from '@angular/fire';
-import { environment } from 'src/environments/environment';
-import { StoreModule } from '@ngrx/store';
-import { appReducer } from './state/app.reducers';
-import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { IonicStorageModule } from '@ionic/storage';
-
-
-export const reducer = {
-  app: appReducer
-};
-
+import { TranslateModule } from '@ngx-translate/core';
+import { TodoService } from './state/todo.service';
 
 @NgModule({
   declarations: [AppComponent],
@@ -31,22 +20,18 @@ export const reducer = {
   imports: [
     BrowserModule,
     IonicStorageModule.forRoot(),
-    StoreModule.forRoot({}),
     HttpClientModule,
     ModalModule,
-    AngularFireModule.initializeApp(environment.firebase),
+    TranslateModule.forRoot({}),
     IonicModule.forRoot(),
     AppRoutingModule,
-    StoreModule.forRoot(reducer),
-    StoreDevtoolsModule.instrument({
-      maxAge: 25 //  Retains last 25 states
-    }),
   ],
   providers: [
+    TodoService,
     StatusBar,
     SplashScreen,
-    { provide: RouteReuseStrategy, useClass: IonicRouteStrategy }
+    { provide: RouteReuseStrategy, useClass: IonicRouteStrategy },
   ],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
 })
 export class AppModule {}
