@@ -1,4 +1,4 @@
-import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+import { Component, Input, Output, EventEmitter } from '@angular/core';
 import * as moment from 'moment';
 
 @Component({
@@ -7,25 +7,23 @@ import * as moment from 'moment';
   styleUrls: ['./list-item.component.scss'],
 })
 export class ListItemComponent {
-
   readonly console = console;
+  _dueDate: string;
+  @Input() set dueDate(dueDate: string) {
+    if (dueDate) {
+      this._dueDate = moment.default(dueDate).format('dddd, MMM Do YYYY');
+    }
+  }
+  get dueDate() {
+    return this._dueDate;
+  }
+
   @Input() checked;
   @Input() taskTitle;
   @Input() additionalInfo;
-  @Input() set dueDate(dueDate: string) {
-    if (dueDate) {
-      this.formattedDateString = moment.default(dueDate).format('dddd, MMM Do YYYY');
-    }
-  }
   @Input() color;
   @Input() listName;
   @Input() colorStyle = 'primary';
-  formattedDateString: string;
-
-
   @Output() checkedChange = new EventEmitter<boolean>();
-  @Output()  divClicked = new EventEmitter<boolean>();
-
-  constructor() { }
-
+  @Output() itemChanged = new EventEmitter<'trash' | 'edit'>();
 }
