@@ -1,13 +1,3 @@
 export const environment = {
-  production: true,
-  firebase: {
-    apiKey: 'AIzaSyC6d6KwAKHESLjZiQJqVgAZCX74Xty_3rE',
-    authDomain: 'todora-app.firebaseapp.com',
-    databaseURL: 'https://todora-app.firebaseio.com',
-    projectId: 'todora-app',
-    storageBucket: 'todora-app.appspot.com',
-    messagingSenderId: '179733328288',
-    appId: '1:179733328288:web:77c6651e0e9b27adac34c5',
-    measurementId: 'G-707LLCFNJJ'
-  }
+  production: true
 };
