@@ -3,17 +3,7 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
-  firebase: {
-    apiKey: 'AIzaSyC6d6KwAKHESLjZiQJqVgAZCX74Xty_3rE',
-    authDomain: 'todora-app.firebaseapp.com',
-    databaseURL: 'https://todora-app.firebaseio.com',
-    projectId: 'todora-app',
-    storageBucket: 'todora-app.appspot.com',
-    messagingSenderId: '179733328288',
-    appId: '1:179733328288:web:77c6651e0e9b27adac34c5',
-    measurementId: 'G-707LLCFNJJ'
-  }
+  production: false
 };
 
 /*
