@@ -1,11 +1,9 @@
-import { AppTypes } from '../../../app/state/app.actions';
 import { ColorService } from '../../services/color/color.service';
-import { IList } from '../../../app/state/app.service';
 import { Component, OnInit } from '@angular/core';
 import { v4 as uuidv4 } from 'uuid';
 import { ModalController } from '@ionic/angular';
-import { Store } from '@ngrx/store';
-
+import { TodoService } from 'src/app/state/todo.service';
+import { IList } from 'src/app/state/todo.interface';
 
 @Component({
   selector: 'app-add-list-modal',
@@ -13,33 +11,32 @@ import { Store } from '@ngrx/store';
   styleUrls: ['./add-list.modal.scss'],
 })
 export class AddListModal implements OnInit {
-
   list: IList = {
     id: uuidv4(),
     listName: null,
     count: 0,
     completed: 0,
-    color: this.colorService.getRandomColor()
+    color: this.colorService.getRandomColor(),
   };
 
-  constructor(private modalController: ModalController,
-              private store: Store<any>,
-              private colorService: ColorService) { }
+  constructor(
+    private modalController: ModalController,
+    private todoService: TodoService,
+    private colorService: ColorService
+  ) {}
 
-  dimiss() {
+  dismiss() {
     this.modalController.dismiss({
-      dismissed: true
+      dismissed: true,
     });
   }
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
   onSave(inputForm) {
     if (inputForm.form.valid) {
-      this.store.dispatch({ type: AppTypes.ADD_LIST, newList: this.list });
-      this.dimiss();
+      this.todoService.addListItem(this.list);
+      this.dismiss();
     }
-
   }
 }
