@@ -1,13 +1,10 @@
-import { getList } from 'src/app/state/app.selector';
-import { AppTypes } from 'src/app/state/app.actions';
-import { AppService, ITodo } from 'src/app/state/app.service';
 import { Observable } from 'rxjs';
 import { ModalController } from '@ionic/angular';
-import { Component, OnInit, OnChanges, SimpleChanges, Input } from '@angular/core';
-import * as uuid from 'uuid';
-import { Store } from '@ngrx/store';
+import { Component, OnInit, Input } from '@angular/core';
 import { v4 as uuidv4 } from 'uuid';
-
+import { TodoService } from 'src/app/state/todo.service';
+import { NgForm } from '@angular/forms';
+import { ITodo } from 'src/app/state/todo.interface';
 
 @Component({
   selector: 'app-add-task',
@@ -15,7 +12,6 @@ import { v4 as uuidv4 } from 'uuid';
   styleUrls: ['./add-task.modal.scss'],
 })
 export class AddTaskModal implements OnInit {
-
   @Input() listValue;
 
   list: Observable<any>;
@@ -26,32 +22,29 @@ export class AddTaskModal implements OnInit {
     listId: null,
     extraDetails: null,
     status: false,
-    dueDate: null
+    dueDate: null,
   };
 
   constructor(
     private modalController: ModalController,
-    private store: Store<any>,
-    private appService: AppService) { }
+    private todoService: TodoService
+  ) {}
 
   ngOnInit() {
-    this.list = this.store.select(getList);
+    this.list = this.todoService.getListItems();
     this.todo.listId = this.listValue;
-    console.log(this.listValue);
   }
 
-  dimiss() {
+  dismiss() {
     this.modalController.dismiss({
-      dismissed: true
+      dismissed: true,
     });
   }
 
-  onSave(inputForm) {
-    console.log(inputForm);
+  onSave(inputForm: NgForm) {
     if (inputForm.form.valid) {
-      this.store.dispatch({ type: AppTypes.ADD_TODO_FULFILLED, todo: [this.todo]});
-      this.dimiss();
-      }
+      this.todoService.addTodoItem(this.todo);
+      this.dismiss();
     }
-
+  }
 }
